many changes to pass mut.py tests with MySQL 8.0 Still some tests fail, status in the mysql-test/test_status.ods spreadsheet.
Note that "form" support was removed in MySQL 8.0;
there are a couple of (acknowledged) bugs in MySQL in the
handling of information_schema tables that are related to
SHOW CREATE ... type statements, the most troubling of which
is in the handling of strings with internal backticks.

So the tests that fail as a result have been commented out.